CSSTUDIO-3171: Embed images
